Skal ikke ha med devdependencies i docker-bygget
