+++ .github/workflows/build_n_deploy_dev.yaml
@@ -23,6 +23,10 @@ jobs:
           NODE_AUTH_TOKEN: ${{ secrets.READER_TOKEN }}
       - name: Yarn build
         run: yarn build
+      - name: Prune dev dependencies
+        run: yarn --prod --prefer-offline --frozen-lockfile
+        env:
+          NODE_AUTH_TOKEN: ${{ secrets.READER_TOKEN }}
       - name: Push docker image to GAR and sign image
         uses: nais/docker-build-push@v0
         id: docker-build-push
